Most of the way implementation of alarm code.
I believe rn it panics when the idle thread tries to
reacquire a lock. I'll have to look into that further.

Most of the code changes are in the timer_inturrupt function,
as well as a few additional functions and variables to facilitate
saving a thread with it's desired wakeup time in a list of
sleeping threads.